requirements.txt
# Core
pyserial

# Remote Approver Agent
discord.py
PyAutoGUI
opencv-python
pillow  # PyAutoGUI dependency, explicit is better
